pyproject.toml
[project]
name = "rainbow-dqn"
version = "0.1.0"
description = "Comparison of RL pixel-based algorithms on the DeepMind Control Suite"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "gymnasium[mujoco]>=1.2.3",
    "imageio[ffmpeg]>=2.37.2",
    "tensorboard>=2.18.0",
    "optuna>=4.4.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.67.3",
    "setuptools<82",
    "opencv-python>=4.13.0.92",
    "plotly>=6.6.0",
    "scikit-learn>=1.8.0",
    "optuna-dashboard>=0.20.0",
    "tensordict>=0.5.0",
    "torchrl>=0.5.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]
torchaudio = [{ index = "pytorch-cu126" }]

